Fix Strong Reference Cycle in DetailTorrentVC
INCOMPLETE COMMIT

I started the Torrent Options Section however I have not gotten around to finishing it. I need to ensure that the table view does not reload if the keyboard is active. I also need to add the remaining rows and a button to upload the config to the server