Fix Almofire ‘cancelled’ & add request retrier
I removed the static validate method that was causing the ‘cancelled’ error. I believe the issue was that the session manager was being deallocated before the request completed. I also implemented a request retrier which seems to fix the issue of requests failing while the application is in the background.

Also slight improvement to MainTableView Status Header.